refactor(xaa): move the state machine into the SDK; Inspector via shim
Relocate createXAAStateMachine (~1940 lines) to sdk/src/xaa/state-machines/,
exported from the browser entry, and add runXaaStateMachine (advance until
complete / stop step / predicate). Move the engine-consumed
NEGATIVE_TEST_MODE_DETAILS table to the SDK; shared/xaa.ts is now a pure
re-export. The inspector's client state-machine.ts becomes a thin re-export
shim, so its adapter and the whole UI are unchanged. The moved 46-test suite
passes verbatim (behavior preserved); the browser guard confirms the engine is
node-free. Swapping the engine's internal discovery/mcp-init onto the PR3
helpers is deferred pending candidate-equivalence verification.
